pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "enzyme-resistance"
version = "0.1.0"
description = "Electrical Resistance Model of Mutation Propagation in Enzymes"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Hari Jayanthan"},
]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "networkx>=3.0",
    "biopython>=1.81",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "requests>=2.28",
    "matplotlib>=3.7",
    "tqdm>=4.65",
]

[project.optional-dependencies]
hf = ["datasets>=2.14", "pyarrow>=12.0"]
dev = ["pytest>=7.0", "pytest-cov"]

[project.scripts]
enzyme-resistance = "enzyme_resistance.cli:main"

[tool.setuptools.packages.find]
include = ["enzyme_resistance*"]
